fix(matching): block alignment
blocks with text that exceeds with and no breaking spaces (e.g. `https://smartycards.cla.umn.edu`) causes problems with with the flex container. This sets the alignment of the block to be to the left, but with an auto margin, which should center it if the block width is smaller than the card width.