we don't need soft deletes for deck memberships
adding soft deletes on deck memberships introduces complexity. Because there's a uniqueness constrain on the `deck_memberships` table, re-adding a member would mean that we need to restore the previous membership OR we need to work around the uniqueness constraint (perhaps by adding an additional constraint of `deleted_at`).

In practice though, we don't need this soft delete. The user's practice data is directly associated with the `user_id` and the `deck_id`, not a `deck_memberships.id`. So if we removing a deck membership, any practice data would not be swept up in the cascade.