fix: redirect to intended path after login
If user login is required, the client wasn't redirecting to the intended path.

Laravel will track the intended path in a session if auth is required, and then redirect back to the intended path after successful auth.

However, on the client side, we were redirecting the client directly to the login page if they weren't authed, so Laravel couldn't see the intended path.

Tracking the intended path on the client side seems complicated – we're likely need to put data in local storage, so that when the app reloads on return we'd be able to complete the request.

A simpler approach seems to be get the client out the way and let Laravel handle auth.

That is optimistically navigate to a page, and if we get a 401 when doing any api calls or navigating to a page that requires auth, we can reload, which triggers a reauthentication, and lets laravel set the intended path as the current page.

This:
- wraps the fallback route in auth middleware
- permits users access to the homepage without auth
- removes the LoginPage
- updates the Sign In link to just go to the Deck Index page, which assumes they're signed in. The client-side will check if the user is authed, and when it's not the client will try loading the intended path, which will trigger Laravel's auth middleware.